Create base model
