+++ index.js
@@ -1,5 +1,6 @@
 'use strict';
-const Build = require('./lib/buildmodel');
-const Pipeline = require('./lib/pipelinemodel');
+const Build = require('./lib/build');
+const Job = require('./lib/job');
+const Pipeline = require('./lib/pipeline');
 
-module.exports = { Build, Pipeline };
+module.exports = { Build, Job, Pipeline };
