+++ lib/build.js
@@ -3,7 +3,8 @@ const async = require('async');
 const Executor = require('screwdriver-executor-k8s');
 const hashr = require('screwdriver-hashr');
 const hoek = require('hoek');
-const PipelineModel = require('./pipelinemodel');
+const BaseModel = require('./base');
+const PipelineModel = require('./pipeline');
 
 /**
  * Given a Job ID, look up the associated Pipeline ID
@@ -49,7 +50,7 @@ function fetchScmUrl(datastore, params, callback) {
     });
 }
 
-const BuildModel = class {
+class BuildModel extends BaseModel {
     /**
      * Construct a BuildModel object
      * @method constructor
@@ -57,7 +58,7 @@ const BuildModel = class {
      * @param  {Object}    [executorOptions] Options to configure the executor-k8s module directly
      */
     constructor(datastore) {
-        this.datastore = datastore;
+        super(datastore);
         this.executor = new Executor({});
         this.table = 'builds';
     }
@@ -132,61 +133,6 @@ const BuildModel = class {
     stream(config, response) {
         return this.executor.stream(config, response);
     }
-
-    /**
-     * Get a build based on id
-     * @param  {String}   id                The id of the record to retrieve
-     * @return {Function} callback          fn(err, result) where result is the build with the specific id
-     */
-    get(id, callback) {
-        const config = {
-            table: this.table,
-            params: {
-                id
-            }
-        };
-
-        return this.datastore.get(config, callback);
-    }
-
-    /**
-     * List builds with pagination
-     * @param  {Object}   paginate           Config object
-     * @param  {Number}   paginate.count     Number of items per page
-     * @param  {Number}   paginate.page      Specific page of the set to return
-     * @return {Function} callback           fn(err, result) where result is an array of builds
-     */
-    list(paginate, callback) {
-        const config = {
-            table: this.table,
-            params: {},
-            paginate: {
-                count: paginate.count,
-                page: paginate.page
-            }
-        };
-
-        return this.datastore.scan(config, callback);
-    }
-
-    /**
-     * Update a build
-     * @param  {Object}    config         Config object
-     * @param  {String}    config.id      The id of the record to retrieve
-     * @param  {Object}    config.data    The new data object to update with
-     * @return {Function}  callback       fn(err, result) where result is the new build object
-     */
-    update(config, callback) {
-        const datastoreConfig = {
-            table: this.table,
-            params: {
-                id: config.id,
-                data: config.data
-            }
-        };
-
-        return this.datastore.update(datastoreConfig, callback);
-    }
-};
+}
 
 module.exports = BuildModel;
