+++ lib/pipeline.js
@@ -1,16 +1,17 @@
 'use strict';
 const hashr = require('screwdriver-hashr');
 const hoek = require('hoek');
-const JobModel = require('./jobmodel');
+const BaseModel = require('./base');
+const JobModel = require('./job');
 
-const PipelineModel = class {
+class PipelineModel extends BaseModel {
     /**
      * Construct a PipelineModel object
      * @method constructor
      * @param  {Object}    datastore         Object that will perform operations on the datastore
      */
     constructor(datastore) {
-        this.datastore = datastore;
+        super(datastore);
         this.table = 'pipelines';
     }
 
@@ -75,61 +76,6 @@ const PipelineModel = class {
             return jobModel.create(jobConfig, callback);
         });
     }
-
-    /**
-    * Get a pipeline based on id
-    * @param  {String}   id                The id of the record to retrieve
-    * @return {Function} callback          fn(err, result) where result is the pipeline with the specific id
-    */
-    get(id, callback) {
-        const config = {
-            table: this.table,
-            params: {
-                id
-            }
-        };
-
-        return this.datastore.get(config, callback);
-    }
-
-   /**
-     * List pipelines with pagination
-     * @param  {Object}   paginate           Config object
-     * @param  {Number}   paginate.count     Number of items per page
-     * @param  {Number}   paginate.page      Specific page of the set to return
-     * @return {Function} callback           fn(err, result) where result is an array of pipelines
-     */
-    list(paginate, callback) {
-        const config = {
-            table: this.table,
-            params: {},
-            paginate: {
-                count: paginate.count,
-                page: paginate.page
-            }
-        };
-
-        return this.datastore.scan(config, callback);
-    }
-
-    /**
-    * Update a pipeline
-    * @param  {Object}    config         Config object
-    * @param  {String}    config.id      The id of the record to retrieve
-    * @param  {Object}    config.data    The new data object to update with
-    * @return {Function}  callback       fn(err, result) where result is the new pipeline object
-    */
-    update(config, callback) {
-        const datastoreConfig = {
-            table: this.table,
-            params: {
-                id: config.id,
-                data: config.data
-            }
-        };
-
-        return this.datastore.update(datastoreConfig, callback);
-    }
-};
+}
 
 module.exports = PipelineModel;
