--- test/index.test.js
@@ -1,8 +1,0 @@
-'use strict';
-const assert = require('chai').assert;
-
-describe('index test', () => {
-    it('fails', () => {
-        assert.isTrue(true);
-    });
-});
